MINT-1788 Decouple userMonitorService from core
Move UserMonitorService and corresponding actions
to @chronobank/login-ui package. Start service after session is created
