+++ packages/login-ui/redux/thunks.js
@@ -23,6 +23,7 @@ import {
 } from '@chronobank/core/redux/persistAccount/constants'
 import * as NetworkThunks from '@chronobank/login/redux/network/thunks'
 import * as SessionThunks from '@chronobank/core/redux/session/thunks'
+import { modalsOpen, modalsClose } from '@chronobank/core/redux/modals/actions'
 import * as PersistAccountActions from '@chronobank/core/redux/persistAccount/actions'
 import PublicBackendProvider from '@chronobank/login/network/PublicBackendProvider'
 import { SignerMemoryModel } from '@chronobank/core/models'
@@ -35,6 +36,7 @@ import {
   FORM_LOGIN_PAGE_FIELD_SUCCESS_MESSAGE,
   FORM_FOOTER_EMAIL_SUBSCRIPTION,
 } from './constants'
+import userMonitorService from './userMonitorService'
 
 /*
  * Thunk dispatched by "" screen.
@@ -122,6 +124,9 @@ export const onSubmitLoginForm = (password) => async (dispatch, getState) => {
         selectedNetworkId,
       ))
       localStorage.createSession(selectedAccount, selectedProviderId, selectedNetworkId)
+
+      dispatch(watchInitUserMonitor())
+
       const defaultURL = await dispatch(SessionThunks.login(selectedAccount))
       dispatch(replace(localStorage.getLastURL() || defaultURL))
     }
@@ -324,4 +329,27 @@ export const onSubmitSubscribeNewsletterFail = (errors, submitErrors) =>
 export const onSubmitLoginFormFail = (errors, submitErrors) => (dispatch) => {
   dispatch(stopSubmit(FORM_LOGIN_PAGE, submitErrors && submitErrors.errors))
   dispatch(NetworkActions.networkResetLoginSubmitting())
+}
+
+export const startUserMonitorAndCloseModals = () => (dispatch) => {
+  userMonitorService.start()
+  dispatch(modalsClose())
+}
+
+export const stopUserMonitor = () => () => {
+  userMonitorService.stop()
+}
+
+export const removeWatchersUserMonitor = () => () => {
+  userMonitorService
+    .removeAllListeners('active')
+    .stop()
+}
+
+export const watchInitUserMonitor = () => (dispatch) => {
+  userMonitorService
+    .on('active', () => dispatch(modalsOpen({
+      componentName: 'UserActiveDialog',
+    })))
+    .start()
 }
