+++ src/components/dialogs/UserActiveDialog/UserActiveDialog.jsx
@@ -7,11 +7,11 @@ import PropTypes from 'prop-types'
 import Button from 'components/common/ui/Button/Button'
 import React, { PureComponent } from 'react'
 import { Translate } from 'react-redux-i18n'
-import userMonitorService from 'user/monitorService'
+import userMonitorService from '@chronobank/login-ui/redux/userMonitorService'
 import { connect } from 'react-redux'
 import { logoutAndNavigateToRoot } from 'redux/ui/thunks'
 import { modalsClose } from '@chronobank/core/redux/modals/actions'
-import { stopUserMonitor } from 'redux/ui/actions'
+import { stopUserMonitor } from '@chronobank/login-ui/redux/thunks'
 import ModalDialog from 'components/dialogs/ModalDialog'
 import Timer from 'components/common/Timer/Timer'
