+++ src/redux/ui/actions.js
@@ -7,32 +7,7 @@ import moment from 'moment'
 import { setLocale } from 'react-redux-i18n'
 import localStorage from 'utils/LocalStorage'
 import ipfs from '@chronobank/core/utils/IPFS'
-import userMonitorService from 'user/monitorService'
-import { modalsOpen, modalsClose } from '@chronobank/core/redux/modals/actions'
 import { CHANGE_WALLET_VIEW } from './constants'
-
-export const startUserMonitorAndCloseModals = () => (dispatch) => {
-  userMonitorService.start()
-  dispatch(modalsClose())
-}
-
-export const stopUserMonitor = () => () => {
-  userMonitorService.stop()
-}
-
-export const removeWatchersUserMonitor = () => () => {
-  userMonitorService
-    .removeAllListeners('active')
-    .stop()
-}
-
-export const watchInitUserMonitor = () => (dispatch) => {
-  userMonitorService
-    .on('active', () => dispatch(modalsOpen({
-      componentName: 'UserActiveDialog',
-    })))
-    .start()
-}
 
 export const changeMomentLocale = (locale) => (dispatch) => {
   moment.locale(locale)
